+++ plugins/psv-extractor/src/index.ts
@@ -9,6 +9,7 @@ export const PSVExtractor = (options?: {
   transform?: (value: any) => Flatfile.CellValueUnion
   chunkSize?: number
   parallel?: number
+  debug?: boolean
 }) => {
   console.log(
     'This plugin is deprecated. PSV extraction is now natively supported by the Flatfile Platform.'