+++ plugins/record-hook/src/RecordHook.ts
@@ -1,16 +1,35 @@
-import { FlatfileEvent } from '@flatfile/listener'
-import { FlatfileRecord, FlatfileRecords } from '@flatfile/hooks'
 import { Record_, Records } from '@flatfile/api/api'
-import { RecordTranslater } from './record.translater'
+import { FlatfileRecord, FlatfileRecords } from '@flatfile/hooks'
+import { FlatfileEvent } from '@flatfile/listener'
 import { asyncBatch } from '@flatfile/util-common'
+import { Effect } from 'effect'
+import { RecordTranslater } from './record.translater'
 
 export const RecordHook = async (
   event: FlatfileEvent,
-  handler: (record: FlatfileRecord, event?: FlatfileEvent) => any | Promise<any>
+  handler: (
+    record: FlatfileRecord,
+    event?: FlatfileEvent
+  ) => any | Promise<any>,
+  options: { concurrency?: number; debug?: boolean } = {}
 ) => {
-  return BulkRecordHook(event, async (records, event) => {
-    return records.map((record) => handler(record, event))
-  })
+  const { concurrency } = { concurrency: 10, ...options }
+  return BulkRecordHook(
+    event,
+    async (records, event) => {
+      const handlers = await records.map((record: FlatfileRecord) =>
+        Effect.promise(async () => {
+          await handler(record, event)
+        })
+      )
+      return Effect.runPromise(
+        Effect.all(handlers, {
+          concurrency,
+        })
+      )
+    },
+    options
+  )
 }
 
 export const BulkRecordHook = async (
@@ -19,7 +38,7 @@ export const BulkRecordHook = async (
     records: FlatfileRecord[],
     event?: FlatfileEvent
   ) => any | Promise<any>,
-  options: { chunkSize?: number; parallel?: number } = {}
+  options: { chunkSize?: number; parallel?: number; debug?: boolean } = {}
 ) => {
   try {
     const records = await event.cache.init<Records>(