+++ plugins/xml-extractor/src/index.ts
@@ -1,5 +1,5 @@
-import { parseBuffer } from './parser'
 import { Extractor } from '@flatfile/util-extractor'
+import { parseBuffer } from './parser'
 
 export const XMLExtractor = (options?: {
   separator?: string
@@ -7,6 +7,7 @@ export const XMLExtractor = (options?: {
   transform?: (row: Record<string, any>) => Record<string, any>
   chunkSize?: number
   parallel?: number
+  debug?: boolean
 }) => {
-  return Extractor('.xml', parseBuffer, options)
+  return Extractor('.xml', 'xml', parseBuffer, options)
 }