+++ plugins/job-handler/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @flatfile/plugin-job-handler
 
+## 0.1.2
+
+### Patch Changes
+
+- 6259301: Introducing the response rejection utility! Webhook egress plugin returns a more usable data structure. Minor bug fixes included.
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [b399623]
+  - @flatfile/util-common@0.2.0
+
 ## 0.1.0
 
 ### Minor Changes