+++ plugins/record-hook/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @flatfile/plugin-record-hook
 
+## 1.1.1
+
+### Patch Changes
+
+- 59a8b18: Bug fix to handle asynchronous handlers
+
+## 1.1.0
+
+### Minor Changes
+
+- b399623: Refactor extractors to handle extraction on job:ready, add debug option, fix asyncBatching
+
+### Patch Changes
+
+- Updated dependencies [b399623]
+  - @flatfile/util-common@0.2.0
+
 ## 1.0.4
 
 ### Patch Changes