+++ plugins/tsv-extractor/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @flatfile/plugin-tsv-extractor
 
+## 1.4.0
+
+### Minor Changes
+
+- b399623: Refactor extractors to handle extraction on job:ready, add debug option, fix asyncBatching
+
+### Patch Changes
+
+- Updated dependencies [b399623]
+  - @flatfile/plugin-delimiter-extractor@0.6.0
+
 ## 1.3.0
 
 ### Minor Changes